Improve how highlighting propagates to child entities
Currently, child entities are always highlighted along with the parent because ctx.globalCompositeOperation is set when rendering the parent texture, but the context is never saved/restored so the setting remains until after we have rendered the children.

Default behaviour of highlight function is now to iterate over child entities and apply the highlight to them directly. This can optionally be disabled so we can control highlighting of children independently if required.

The highlight logic in renderCache was removed, as this was incompatible with the new approach and even in the previous code would have had resulted in a double application of highlighting, leading to undesired effects dependent on the operation, such as absolute blacks becoming transparent.